% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDatanet.R
\docType{package}
\name{CDatanet-package}
\alias{CDatanet}
\alias{CDatanet-package}
\title{CDatanet: Count Data Model With Social Interactions}
\description{
Estimating a count data model with social interactions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ahoundetoungan/CDatanet}
  \item Report bugs at \url{https://github.com/ahoundetoungan/CDatanet/issues}
}

}
\author{
\strong{Maintainer}: Elysée Aristide Houndetoungan \email{ariel92and@gmail.com}

}
