% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{foptimREM_nopeer}
\alias{foptimREM_nopeer}
\title{foptimREM_nopeer: computes the likelihood with the constraint alpha = 0
                  theta does not contain alpha here: just the explanatory
                  variables and sigma.}
\usage{
foptimREM_nopeer(theta, X, w, a, A, K, N, R, y)
}
\description{
foptimREM_nopeer: computes the likelihood with the constraint alpha = 0
                  theta does not contain alpha here: just the explanatory
                  variables and sigma.
}
