% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{foptimREM_nopeervec}
\alias{foptimREM_nopeervec}
\title{foptimREM_nopeervec: the vectorized form of foptimREM_nopeer
                     returns the likelihood for each individuals
                      given the patameters}
\usage{
foptimREM_nopeervec(theta, X, w, a, A, K, N, R, y)
}
\description{
foptimREM_nopeervec: the vectorized form of foptimREM_nopeer
                     returns the likelihood for each individuals
                      given the patameters
}
