% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{foptimREMvec}
\alias{foptimREMvec}
\title{foptimREMvec: vectorized form of foptimREM
              returns the likelihood for each
              individuals given the patameters}
\usage{
foptimREMvec(
  P,
  theta,
  X,
  G,
  igroup,
  ngroup,
  w,
  a,
  A,
  K,
  N,
  R,
  y,
  tol = 1e-13,
  maxit = 1000L
)
}
\description{
foptimREMvec: vectorized form of foptimREM
              returns the likelihood for each
              individuals given the patameters
}
