% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdnet.R
\name{meffects.cdnet}
\alias{meffects.cdnet}
\title{Computing marginal effects for count data models with social interactions under rational expectations}
\usage{
meffects.cdnet(object, Glist, data, list.index = list(), S = NULL, ncores = 1L)
}
\arguments{
\item{object}{an object of class \code{cdnet} or \code{summary.cdnet}, as returned by the function \code{\link{cdnet}} or \code{\link{summary.cdnet}}.}

\item{Glist}{adjacency matrix. For networks consisting of multiple subnets, \code{Glist} can be a list of subnets with the \code{m}-th element being an \code{ns*ns} adjacency matrix, where \code{ns} is the number of nodes in the \code{m}-th subnet.
For heterogenous peer effects (e.g., boy-boy, boy-girl friendship effects), the \code{m}-th element can be a list of many \code{ns*ns} adjacency matrices corresponding to the different network specifications (see Houndetoungan, 2024).
For heterogeneous peer effects in the case of a single large network, \code{Glist} must be a one-item list. This item must be a list of many specifications of large networks.}

\item{data}{an optional data frame, list or environment (or object coercible by \link[base]{as.data.frame} to a data frame) containing the variables
in the model. If not found in data, the variables are taken from \code{environment(formula)}, typically the environment from which \code{summary.cdnet} is called.}

\item{list.index}{list of observation indices to use to calculate the average marginal effect for each variable (see the example of the function \code{\link{cdnet}}).}

\item{S}{he number of simulations to use to calculate standard errors. Only marginal effects are calculated when \code{S = NULL}.}

\item{ncores}{ncores the number of cores to use for simulations. A large number of cores should make the process faster.}
}
\description{
\code{meffects.cdnet} computes and prints marginal effects for the class \code{cdnet} as returned by the function \link{cdnet}.
Standard errors for the marginal effects are computed by simulation following Krinsky and Robb (1989).
}
\references{
Krinsky, I., & Robb, A. L. (1986). On approximating the statistical properties of elasticities. The review of economics and statistics, 715-719, \doi{10.2307/1924536}.
}
