% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdnet.R
\name{print.meffects.cdnet}
\alias{print.meffects.cdnet}
\alias{summary.meffects.cdnet}
\title{Printing marginal effects for count data models with social interactions under rational expectations}
\usage{
\method{print}{meffects.cdnet}(x, ...)

\method{summary}{meffects.cdnet}(object, ...)
}
\arguments{
\item{x}{an object of class \code{meffects.cdnet} as return by the function \code{\link{meffects.cdnet}} or \code{\link{print.meffects.cdnet}}.}

\item{...}{further arguments passed to or from other methods.}

\item{object}{an object of class \code{meffects.cdnet} as return by the function \code{\link{meffects.cdnet}}.}
}
\description{
Summary and print methods for the class \code{meffects.cdnet} as returned by the function \link{meffects.cdnet}.
}
