% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdnet.R
\name{print.simcdEy}
\alias{print.simcdEy}
\alias{summary.simcdEy}
\alias{print.summary.simcdEy}
\title{Printing the average expected outcomes for count data models with social interactions}
\usage{
\method{print}{simcdEy}(x, ...)

\method{summary}{simcdEy}(object, ...)

\method{print}{summary.simcdEy}(x, ...)
}
\arguments{
\item{x}{an object of class \code{summary.simcdEy}, output of the function \code{\link{summary.simcdEy}}
or class \code{simcdEy}, output of the function \code{\link{simcdEy}}.}

\item{...}{further arguments passed to or from other methods.}

\item{object}{an object of class \code{simcdEy}, output of the function \code{\link{simcdEy}}.}
}
\value{
A list of the same objects in \code{object}.
}
\description{
Summary and print methods for the class \code{simcdEy} as returned by the function \link{simcdEy}.
}
