% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdnet.R
\name{simcdpar}
\alias{simcdpar}
\title{Simulating the distribution of the NPL estimator for count data models with social interactions}
\usage{
simcdpar(object, n.simu = 10L, ...)
}
\arguments{
\item{object}{an object of class \code{summary.cdnet}, output of the function \code{\link{summary.cdnet}}
or class \code{cdnet}, output of the function \code{\link{cdnet}}.}

\item{n.simu}{is the number of simu}

\item{...}{further arguments passed to \link[MASS]{mvrnorm}.}
}
\value{
A vector of the same dimension as the parameter estimate from \code{cdnet} if \code{n.simu= 1} or a matrix with each row being one simulation otherwise.
}
\description{
\code{simcdpar} simulate from the distribution of the NPL estimator for count data models with social interactions.
Because constraints are imposed on the parameters, these simulations are performed using the unconstrained parameters,
and the results are transformed into constrained parameters. This ensures that the simulations verify the same constraints
as the parameters.
}
