% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SARestim.R
\name{summary.sar}
\alias{summary.sar}
\alias{print.summary.sar}
\alias{print.sar}
\alias{summary.sars}
\alias{print.summary.sars}
\alias{print.sars}
\title{Summarize SAR Model}
\usage{
\method{summary}{sar}(object, ...)

\method{print}{summary.sar}(x, ...)

\method{print}{sar}(x, ...)

\method{summary}{sars}(object, ...)

\method{print}{summary.sars}(x, ...)

\method{print}{sars}(x, ...)
}
\arguments{
\item{object}{an object of class \code{sar}, output of the function \code{\link{sar}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class \code{summary.sar}, output of the function \code{\link{summary.sar}} or
class \code{sar}, output of the function \code{\link{sar}}.}
}
\value{
A list of the same objects in \code{object}.
}
\description{
Summary and print methods for the class \code{sar} as returned by the function \link{sar}.
}
