% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAR.R
\name{summary.SARML}
\alias{summary.SARML}
\alias{print.summary.SARML}
\alias{print.SARML}
\title{Summarize SAR Model}
\usage{
\method{summary}{SARML}(object, ...)

\method{print}{summary.SARML}(x, ...)

\method{print}{SARML}(x, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary and print methods for the class `SARML` as returned by the function \link{SARML}.
}
