% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SARTestim.R
\name{summary.SARTML}
\alias{summary.SARTML}
\alias{print.summary.SARTML}
\alias{print.SARTML}
\alias{print.summary.SARTMLs}
\title{Summarize SART Model}
\usage{
\method{summary}{SARTML}(object, ...)

\method{print}{summary.SARTML}(x, ...)

\method{print}{SARTML}(x, ...)

\method{print}{summary.SARTMLs}(x, ...)
}
\arguments{
\item{object}{an object of class \code{SARTML}, output of the function \code{\link{SARTML}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class \code{summary.SARTML}, output of the function \code{\link{summary.SARTML}}
or class \code{SARTML}, output of the function \code{\link{SARTML}}.}
}
\value{
A list of the same objects in \code{object}.
}
\description{
Summary and print methods for the class \code{SARTML} as returned by the function \link{SARTML}.
}
