% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDatanet.R
\docType{package}
\name{CDatanet-package}
\alias{CDatanet}
\alias{CDatanet-package}
\title{CDatanet: Count Data Model With Social Interactions}
\description{
Estimating social networks models where the dependent variable is counting.
}
