% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPL.R
\name{summary.CDnetNPL}
\alias{summary.CDnetNPL}
\alias{print.summary.CDnetNPL}
\alias{print.CDnetNPL}
\title{Summarize Count Data Model With Social Interactions}
\usage{
\method{summary}{CDnetNPL}(object, cov.ctr = list(), ...)

\method{print}{summary.CDnetNPL}(x, ...)

\method{print}{CDnetNPL}(x, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary and print methods for the class `CDnetNPL` as returned by the function \link{CDnetNPL}.
}
