% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDestim.R
\name{summary.CDnetNPL}
\alias{summary.CDnetNPL}
\alias{print.summary.CDnetNPL}
\alias{print.CDnetNPL}
\alias{print.summary.CDnetNPLs}
\title{Summarize Count Data Model with Social Interactions}
\usage{
\method{summary}{CDnetNPL}(object, Glist, data, S = 1000L, ...)

\method{print}{summary.CDnetNPL}(x, ...)

\method{print}{CDnetNPL}(x, ...)

\method{print}{summary.CDnetNPLs}(x, ...)
}
\arguments{
\item{object}{an object of class \code{CDnetNPL}, output of the function \code{\link{CDnetNPL}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class \code{summary.CDnetNPL}, output of the function \code{\link{summary.CDnetNPL}},
class \code{summary.CDnetNPLs}, list of outputs of the function \code{\link{summary.CDnetNPL}}
(when the model is estimated many times to control for the endogeneity)
or class \code{CDnetNPL} of the function \code{\link{CDnetNPL}}.}
}
\value{
A list of the same objects in \code{object}.
}
\description{
Summary and print methods for the class \code{CDnetNPL} as returned by the function \link{CDnetNPL}.
}
