% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SART.R
\name{summary.SARTML}
\alias{summary.SARTML}
\alias{print.summary.SARTML}
\alias{print.SARTML}
\title{Summarize SART Model}
\usage{
\method{summary}{SARTML}(object, ...)

\method{print}{summary.SARTML}(x, ...)

\method{print}{SARTML}(x, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary and print methods for the class `SARTML` as returned by the function \link{SARTML}.
}
